+++ src/ray/common/scheduling/cluster_resource_scheduler.cc
@@ -1,4 +1,4 @@
-#include "scheduling_alg.h"
+#include "cluster_resource_scheduler.h"
 using namespace std;
 
 ClusterResourceScheduler::ClusterResourceScheduler(
